+++ .github/workflows/bioscrape.yml
@@ -3,8 +3,6 @@ on: [push]
 jobs:
   test:
     runs-on: ubuntu-latest
-    strategy:
-      python-version: "3.10"
     steps:
       - uses: actions/checkout@v2
       - name: Set up Python v3.10
@@ -22,7 +20,6 @@ jobs:
       - name: install lineage
         run: python setup.py install lineage
   deploy:
-    needs: test
     if: github.ref == 'refs/heads/master'
     runs-on: ${{ matrix.os }}
     strategy:
