+++ .github/workflows/bioscrape.yml
@@ -4,9 +4,9 @@ jobs:
   test:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - name: Set up Python v3.10
-        uses: actions/setup-python@v2
+        uses: actions/setup-python@v3
         with:
           python-version: "3.10"
       - name: Install dependencies
@@ -20,6 +20,7 @@ jobs:
       - name: install lineage
         run: python setup.py install lineage
   deploy:
+    needs: test
     if: github.ref == 'refs/heads/master'
     runs-on: ${{ matrix.os }}
     strategy:
