+++ .github/workflows/bioscrape.yml
@@ -13,12 +13,12 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
-      - name: install bioscrape and lineage
-        run: python setup.py install
-      - name: Test bioscrape with pytest
-        run: pytest
-      - name: install lineage
-        run: python setup.py install lineage
+      # - name: install bioscrape and lineage
+      #   run: python setup.py install
+      # - name: Test bioscrape with pytest
+      #   run: pytest
+      # - name: install lineage
+      #   run: python setup.py install lineage
   deploy:
     needs: test
     if: github.ref == 'refs/heads/master'
