+++ .github/workflows/deploy_bioscrape.yml
@@ -1,0 +1,27 @@
+name: deploy_package
+on:
+  push:
+    branches:
+      - master
+  runs-on: ${{ matrix.os }}
+  strategy:
+    max-parallel: 4
+    matrix:
+      os: [windows-latest, ubuntu-latest, macos-latest]
+      python-version: ["3.7", "3.8", "3.9", "3.10"]
+  steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v2
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
+    - name: install bioscrape and lineage
+      run: python setup.py install
+    - name: Test bioscrape with pytest
+      run: pytest
+    - name: install lineage
+      run: python setup.py install lineage
