+++ .github/workflows/bioscrape.yml
@@ -4,24 +4,23 @@ on:
     branches:
       - '*'
       - '!master'
-  jobs:
-      runs-on: ubuntu-latest
-      steps:
-        - uses: actions/checkout@v3
-        - name: Set up Python v3.10
-          uses: actions/setup-python@v3
-          with:
-            python-version: "3.10"
-        - name: Install dependencies
-          run: |
-            python -m pip install --upgrade pip
-            pip install -r requirements.txt
-        # - name: install bioscrape and lineage
-        #   run: python setup.py install
-        # - name: Test bioscrape with pytest
-        #   run: pytest
-        # - name: install lineage
-        #   run: python setup.py install lineage
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - name: Set up Python v3.10
+        uses: actions/setup-python@v3
+        with:
+          python-version: "3.10"
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt
+      # - name: install bioscrape and lineage
+      #   run: python setup.py install
+      # - name: Test bioscrape with pytest
+      #   run: pytest
+      # - name: install lineage
+      #   run: python setup.py install lineage
 
 # on: [push]
 # jobs:
