+++ .github/workflows/bioscrape.yml
@@ -1,69 +1,0 @@
-name: test_package
-on:
-  push:
-    branches:
-      - '*'
-      - '!master'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Python v3.10
-        uses: actions/setup-python@v3
-        with:
-          python-version: "3.10"
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install -r requirements.txt
-      # - name: install bioscrape and lineage
-      #   run: python setup.py install
-      # - name: Test bioscrape with pytest
-      #   run: pytest
-      # - name: install lineage
-      #   run: python setup.py install lineage
-
-# on: [push]
-# jobs:
-#   test:
-#     runs-on: ubuntu-latest
-#     steps:
-#       - uses: actions/checkout@v3
-#       - name: Set up Python v3.10
-#         uses: actions/setup-python@v3
-#         with:
-#           python-version: "3.10"
-#       - name: Install dependencies
-#         run: |
-#           python -m pip install --upgrade pip
-#           pip install -r requirements.txt
-#       # - name: install bioscrape and lineage
-#       #   run: python setup.py install
-#       # - name: Test bioscrape with pytest
-#       #   run: pytest
-#       # - name: install lineage
-#       #   run: python setup.py install lineage
-#   deploy:
-#     needs: test
-#     if: github.ref == 'refs/heads/master'
-#     runs-on: ${{ matrix.os }}
-#     strategy:
-#       max-parallel: 4
-#       matrix:
-#         os: [windows-latest, ubuntu-latest, macos-latest]
-#         python-version: ["3.7", "3.8", "3.9", "3.10"]
-#     steps:
-#       - uses: actions/checkout@v2
-#       - name: Set up Python ${{ matrix.python-version }}
-#         uses: actions/setup-python@v2
-#         with:
-#           python-version: ${{ matrix.python-version }}
-#       - name: Install dependencies
-#         run: |
-#           python -m pip install --upgrade pip
-#           pip install -r requirements.txt
-#       - name: install bioscrape and lineage
-#         run: python setup.py install
-#       - name: Test bioscrape with pytest
-#         run: pytest
-#       - name: install lineage
-#         run: python setup.py install lineage
