+++ .github/workflows/test_bioscrape.yml
@@ -1,15 +1,16 @@
-name: deploy_package
+name: test_package
+name: test_package
 on:
   push:
-    branches: ["master"]
+    branches: ["*","!master"]
 jobs:
   build:
     runs-on: ${{ matrix.os }}
     strategy:
       max-parallel: 4
       matrix:
-        os: [windows-latest, ubuntu-latest, macos-latest]
-        python-version: ["3.8", "3.9", "3.10"]
+        os: [ubuntu-latest]
+        python-version: ["3.10"]
 
     steps:
     - uses: actions/checkout@v3
