+++ .github/workflows/deploy_bioscrape.yml
@@ -2,7 +2,7 @@ name: deploy_package
 on:
   push:
     branches: ["master"]
-  pull-request:
+  pull_request:
     branches: ["master"]  
 jobs:
   build:
