+++ .github/workflows/test_bioscrape.yml
@@ -1,4 +1,4 @@
-name: test_package
+name: test_bioscrape
 on:
   push:
     branches: ["*","!master"]
