+++ paper.bib
@@ -55,6 +55,7 @@
 @InProceedings{ mckinney-proc-scipy-2010,
   author    = { Wes McKinney },
   title     = { Data Structures for Statistical Computing in Python },
+  title     = {10.25080/majora-92bf1922-00a},
   booktitle = { Proceedings of the 9th Python in Science Conference },
   pages     = { 51 - 56 },
   year      = { 2010 },
@@ -76,6 +77,7 @@
   title={Parameter estimation using metaheuristics in systems biology: a comprehensive review},
   author={Sun, Jianyong and Garibaldi, Jonathan M and Hodgman, Charlie},
   journal={IEEE/ACM Transactions on Computational Biology and Bioinformatics (TCBB)},
+  doi={10.1109/tcbb.2011.63},
   volume={9},
   number={1},
   pages={185--202},
@@ -98,6 +100,7 @@
   title={Resource usage and gene circuit performance characterization in a cell-free breadboard},
   author={Siegal-Gaskins, Dan and Tuza, Zoltan A and Kim, Jongmin and Noireaux, Vincent and Murray, Richard M},
   journal={BioRxiv},
+  doi={10.1101/000885},
   pages={000885},
   year={2014},
   publisher={Cold Spring Harbor Laboratory}
@@ -113,7 +116,8 @@
 @article{kondo2010reaction,
   title={Reaction-diffusion model as a framework for understanding biological pattern formation},
   author={Kondo, Shigeru and Miura, Takashi},
-  journal={science},
+  journal={Science},
+  doi={10.1126/science.1179047},
   volume={329},
   number={5999},
   pages={1616--1620},
@@ -124,7 +128,8 @@
 @article{adcock2006molecular,
   title={Molecular dynamics: survey of methods for simulating the activity of proteins},
   author={Adcock, Stewart A and McCammon, J Andrew},
-  journal={Chemical reviews},
+  journal={Chemical Reviews},
+  doi={10.1002/chin.200630297},
   volume={106},
   number={5},
   pages={1589--1615},
@@ -136,6 +141,7 @@
   title={Excluded-Volume Effects in Living Cells},
   author={Gnutt, David and Gao, Mimi and Brylski, Oliver and Heyden, Matthias and Ebbinghaus, Simon},
   journal={Angewandte Chemie International Edition},
+  doi={10.1002/anie.201409847},
   volume={54},
   number={8},
   pages={2548--2551},
@@ -148,6 +154,7 @@
 	title = {Bayesian parameter inference for stochastic biochemical network models using particle Markov chain Monte Carlo},
 	year = {2011},
 	publisher = {Royal Society},
+  doi={10.1098/rsfs.2011.0047},
 	journal = {Interface Focus}
 }
 
@@ -156,6 +163,7 @@
   title={Moment-based inference predicts bimodality in transient gene expression},
   author={Zechner, Christoph and Ruess, Jakob and Krenn, Peter and Pelet, Serge and Peter, Matthias and Lygeros, John and Koeppl, Heinz},
   journal={Proceedings of the National Academy of Sciences},
+  doi={10.1073/pnas.1200161109},
   volume={109},
   number={21},
   pages={8340--8345},
@@ -177,6 +185,7 @@
 	volume = {11},
 	number = {2},
 	journal = {Nature Methods},
+  doi = {10.1038/nmeth.2794},
 	author = {Zechner, Christoph and Unger, Michael and Pelet, Serge and Peter, Matthias and Koeppl, Heinz},
 	year = {2014},
[…]
