### 2022-11-29 b093fcf
add characterization paper citation
