+++ paper.bib
@@ -660,6 +660,13 @@ number = {6},
 pages = {387-397},
 year = {2014}
 }
+@article{pandey2022characterization,
+  title={Characterization of Integrase and Excisionase Activity in Cell-free Protein Expression System Using a Modeling and Analysis Pipeline},
+  author={Pandey, Ayush and Rodriguez, Makena L and Poole, William and Murray, Richard M},
+  journal={bioRxiv},
+  year={2022},
+  publisher={Cold Spring Harbor Laboratory}
+}
 
 @article{riboj,
   title={Ribozyme-based insulator parts buffer synthetic circuits from genetic context},
