+++ paper.bib
@@ -413,7 +413,7 @@ Eprint = {arXiv:1605.05517},
   title={Biomolecular Feedback Systems},
   author={Del Vecchio, Domitilla and Murray, Richard M.},
   year={2014},
-  doi = {10.23943/princeton/9780691161532.001.0001}
+  doi = {10.23943/princeton/9780691161532.001.0001},
   publisher={Princeton University Press}
 }
 
