### 2022-11-29 2dcd59c
add joss badge, update CI badge
