+++ .github/workflows/test_bioscrape.yml
@@ -9,7 +9,7 @@ jobs:
       max-parallel: 4
       matrix:
         os: [ubuntu-latest]
-        python-version: ["3.10"]
+        python-version: ["3.7", "3.10"]
 
     steps:
     - uses: actions/checkout@v3
