### 2022-11-29 f295ff8
fix doi characterization paper
