+++ paper.bib
@@ -664,6 +664,7 @@ year = {2014}
   title={Characterization of Integrase and Excisionase Activity in Cell-free Protein Expression System Using a Modeling and Analysis Pipeline},
   author={Pandey, Ayush and Rodriguez, Makena L and Poole, William and Murray, Richard M},
   journal={bioRxiv},
+  doi={10.1101/2022.10.05.511053},
   year={2022},
   publisher={Cold Spring Harbor Laboratory}
 }
