### 2022-11-29 516f5b1
Repo cleanup + Upgrade testing CI env to Python v3.10 (#146)
* Bioscrape v1.2.0 (#141)

* remove 3.6 testing in favor of 3.10 testing
* add quotes around 3.10 for CI 
* remove old trash files: temp.pck, simbio.m etc
* fix compatibility spelling
* fix python capitalization
* try branch specific actions CI
* rename CIs to test_bioscrape and deploy_bioscrape
* add joss badge, update CI badge
