+++ README.md
@@ -1,0 +1,109 @@
+# Bioscrape &mdash; Biological Stochastic Simulation of Single Cell Reactions and Parameter Estimation
+## Python toolbox to simulate, analyze, and learn biological system models
+
+[![Build Status](https://github.com/biocircuits/bioscrape/actions/workflows/deploy_bioscrape.yml/badge.svg)](https://github.com/biocircuits/bioscrape/actions/workflows/deploy_bioscrape.yml)
+[![PyPI version](https://badge.fury.io/py/bioscrape.svg)](https://badge.fury.io/py/bioscrape)
+[![status](https://joss.theoj.org/papers/5935db21ffab5b33069d05b7adbdf094/status.svg)](https://joss.theoj.org/papers/5935db21ffab5b33069d05b7adbdf094)
+
+* Getting started with Bioscrape: [![Bioscrape Core](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/biocircuits/bioscrape/blob/colab-ipynb/examples/Basic%20Examples%20-%20START%20HERE.ipynb#scrollTo=Jmm8mTPfhMMS)
+
+* Bioscrape analysis features: [![Bioscrape Sensitivity Analysis](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/biocircuits/bioscrape/blob/colab-ipynb/examples/Sensitivity%20Analysis%20using%20Bioscrape.ipynb#scrollTo=Gu1r4H4ti_z7)
+
+* Parameter inference with Bioscrape: [![Bioscrape Inference](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/biocircuits/bioscrape/blob/colab-ipynb/inference%20examples/Bioscrape%20Inference%20-%20Getting%20Started.ipynb#scrollTo=yvYVliBgjyzF)
+
+Bioscrape is a Systems Biology Markup Language (SBML) simulator written in Cython for speed and Python compatibility. It can be used for deterministic, stochastic, or single cell simulation and also has parameter inference capabilities.
+
+- **Mailing list:** [SBTools Google Group](https://groups.google.com/g/sbtools/) Email: sbtools@googlegroups.com
+- **Source:** https://github.com/biocircuits/bioscrape
+- **Preprint:** - [Fast and flexible simulation and parameter estimation for synthetic biology using bioscrape](https://www.biorxiv.org/content/10.1101/121152v3)
+- **Bug reports:** https://github.com/biocircuits/bioscrape/issues
+- **Slack** Join the #bioscrape channel on SBTools slack: Ask on the public SBTools Google group to be added or send a message to one of the maintainers. 
+
+# Example 1: Simulating an SBML
+
+Bioscrape allows for deterministic and stochastic simulation of SBML models:
+
+```python
+from bioscrape.types import Model
+# Load an SBML file repressilator.xml 
+# (you can find this file in `examples/models` directory)
+M = Model(sbml_filename = 'repressilator_sbml.xml')
+# Simulate the model
+from bioscrape.simulator import py_simulate_model
+import numpy as np
+tp = np.linspace(0,256,100)
+result = py_simulate_model(timepoints=tp, Model=M, stochastic=True)
+# Plot the simulation result (the result is a Pandas dataframe)
+import matplotlib.pyplot as plt
+plt.plot(tp, result['X'])
+```
+
+# Example 2: Run Bayesian inference with Bioscrape 
+
[…]
