+++ README.rst
@@ -1,15 +1,0 @@
-bioscrape - Biological Stochastic Simulation of Single Cell Reactions and Parameter Estimation
-==============================================================================================
-
-|Build Status|
-
-Bioscrape is a Chemical Reaction Network Simulator written in cython for speed and python compatability. It can be used for deterministic, stochastic, or single cell simulation and also has parameter inference capabilities.
-
-To install Bioscrape, in your terminal run the command:
-   
-   pip install bioscrape
-
-Please note that Bioscrape is a cython extension module and requires a C++ compiler to be set up on your computer for installation. For more information, please visit https://github.com/biocircuits/bioscrape/wiki.
-
-.. |Build Status| image:: https://github.com/biocircuits/bioscrape/actions/workflows/bioscrape.yml/badge.svg
-   :target: https://github.com/biocircuits/bioscrape
