### 2022-11-30 063c593
Inference pickling (#145)
* InferenceSetup pickles

* added __getstate__ and __setstate__ methods with tests

* Allow keyword 'pool' to be used with emcee

Note: this is actually not faster in cases where numpy is automatically threading on multiple cores (which is the default behavior on MacOS and Linux).

Co-authored-by: Ayush Pandey <ayush.9.pandey@gmail.com>
Co-authored-by: William Poole <wpoole@caltech.edu>
