### 2022-11-30 e1c5559
Dev to Master for v1.2 (#148)
Changes summary:
* update badges
* add corner dependency
* update version

* Repo cleanup
* Upgrade testing CI env to Python v3.10 (#146)

* Inference pickling (#145)

Co-authored-by: Ayush Pandey <ayush.9.pandey@gmail.com>
Co-authored-by: William Poole <wpoole@caltech.edu>
