### 2022-11-30 f2649eb
add setup.cfg for pypi
