+++ setup.cfg
@@ -1,0 +1,4 @@
+[metadata]
+long_description_content_type = text/markdown
+[aliases]
+test=pytest
