### 2022-12-10 #4961 @csoneson
👋🏻 @arghdos, @Farnazmdi, @espottesmith - would two of you be interested in reviewing this submission for JOSS?

> Fast and flexible simulation and parameter estimation for synthetic biology using bioscrape
