### 2022-12-10 #4961 @csoneson
> I tried above, but it didn't work. Maybe it was because I didn't start the message with that command.

@ayush9pandey - yes, the editorialbot commands need to be given first in the message.
