### 2022-12-16 #4961 @csoneson
👋🏻 @smsaladi, @ayaanhossain - would you be interested in reviewing this submission for JOSS?

> Fast and flexible simulation and parameter estimation for synthetic biology using bioscrape
