### 2023-01-08 #4961 @csoneson
@jguhlin, @Robaina - would one of you be interested in reviewing this submission for JOSS?

> Fast and flexible simulation and parameter estimation for synthetic biology using bioscrape
