### 2023-01-09 #5057 @editorialbot
[REVIEW]: Fast and flexible simulation and parameter estimation for synthetic biology using bioscrape
**Submitting author:** @ayush9pandey (<a href="http://orcid.org/0000-0003-3590-4459">Ayush Pandey</a>)
**Repository:** https://github.com/biocircuits/bioscrape/
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.2.2
