### 2023-01-14 #5057 @Robaina
Hi @sclamons,

I'm one of the reviewers of this submission (submitted by @ayush9pandey). I'm currently following JOSS guidelines regarding [authorship](https://joss.readthedocs.io/en/latest/review_criteria.html). I see that you have made a substantial amount of commits to the [repo](https://github.com/biocircuits/bioscrape/graphs/contributors) but I cannot map your username to any name in the author list of the submitted manuscript. Are you one of the authors listed in the manuscript? Thank you!
