### 2023-01-14 #5057 @Robaina
Hi @csoneson,

this is my first review in JOSS, it's not entirely clear to me which questions/comments should I post here and which ones in the issues section of the repo. Are there any general guidelines for this? For instance, comments regarding the manuscript should be posted in this thread or in the repo?  Thank you!
