### 2023-01-14 biocircuits/bioscrape#151 @Robaina
JOSS REVIEW - Contributing Guidelines
Hi!

I'm one of the reviewers of this repo at [JOSS](https://github.com/openjournals/joss-reviews/issues/5057). First of all, congratulations on this well-documented and organized repo!

I see that the README includes a section regarding [contribution guidelines](https://github.com/openjournals/joss-reviews/issues/5057). While this section contains minimal info, I think it would be great to extend it a little bit (e,g, setting a developer environment, general ways of contributing, etc). Perhaps even create a separate [CONTRIBUTING.md](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) file.

This is just a suggestion, though.
