### 2023-01-14 #5057 @csoneson
Hi @Robaina - different reviewers have different preferences; in my experience it's often a good idea to post specific improvement suggestions and encountered problems as individual issues in the software repo, where they can be discussed in separate threads and it's easier to keep track of their status, and to keep this review issue more for general "summary" comments or questions. Comments regarding the manuscript can go in either place (as the manuscript also lives in the software repo).
