### 2023-01-14 #5057 @sclamons
Hi Semidán,

I know which paper this is, but I can't actually see the submission (I
graduated and left the lab about a year ago). My name is Sam Clamons, which
you probably could have guessed if it were on the author list! I don't
think I should be, particularly - I made relatively minor contributions
using some unorthodox dev practices that caused me to submit a
disproportionate number of commits.

Thanks for checking! Very much appreciated.
Sam

On Sat, Jan 14, 2023, 2:21 AM Semidán Robaina, Ph.D. <
***@***.***> wrote:

> Hi @sclamons <https://github.com/sclamons>,
>
> I'm one of the reviewers of this submission (submitted by @ayush9pandey
> <https://github.com/ayush9pandey>). I'm currently following JOSS
> guidelines regarding authorship
> <https://joss.readthedocs.io/en/latest/review_criteria.html>. I see that
> you have made a substantial amount of commits to the repo
> <https://github.com/biocircuits/bioscrape/graphs/contributors> but I
> cannot map your username to any name in the author list of the submitted
> manuscript. Are you one of the authors listed in the manuscript? Thank you!
>
