### 2023-01-15 biocircuits/bioscrape#152 @Robaina
JOSS Review - Advance example Notebook broken
Hi!

taking a look at the examples for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/5057) I noticed that the notebook containing the [advanced example](https://github.com/biocircuits/bioscrape/blob/master/examples/Advanced%20Examples%20-%20Direct%20Simulator%20Instantiation.ipynb) is broken. I get an error message: Unexpected token "<" in JSON at position 71312. Most likely easily fixable.
