+++ CONTRIBUTING.md
@@ -1,38 +1,0 @@
-# Contributing to Bioscrape 
-
-Thank you for your interest in contributing to Bioscrape!
-
-In this file you will find detailed instructions on how you can start making contributions to the package. Bioscrape is hosted on the [Biocircuits](https://github.com/biocircuits) organization page on GitHub. Note that the legacy version of the package was hosted on the lead developer, ananswam's account and is not actively maintained. Use the [biocircuits/bioscrape](https://github.com/biocircuits/bioscrape/) for all purposes. For more information on getting started with the package, refer to the README file on the [home page](https://github.com/biocircuits/bioscrape/) and the tutorial style example jupyter notebooks under the [examples](https://github.com/biocircuits/bioscrape/tree/master/examples) directory. The software documentation is available on the Github Wiki page [here](https://github.com/biocircuits/bioscrape/wiki). 
-
-## How to contribute?
-To get started, set up your Bioscrape fork under your account on Github - [raise an issue](https://github.com/biocircuits/bioscrape/issues/new) if you are not sure how to set it up. All contributions to Bioscrape should be made as a Github pull request to the **dev** branch [here](https://github.com/biocircuits/bioscrape/tree/dev). 
-
-### Reporting Bugs/Asking for help
-
-Use the Github issues page on Bioscrape to report a bug or to ask for help with running Bioscrape. The Github issues have labels that you can use so that the issues can be filtered easily:
-
-* If you are unsure where to begin contributing to Bioscrape, you can start by looking through the issues with the label `beginner` or `help-wanted`. A beginner issue usually requires only changing a few lines of code to fix something or add a new enhancement. The `help-wanted` issues are slightly more involved and may require an understanding of the Bioscrape modules.
-
-* If you have a particular feature idea in mind, feel free to suggest that as an `enhancement` or a `feature-request` tagged issue. If you would like to get in touch with the developers working on the package to discuss your contribution ideas, you can also join our Slack channel (details at the end of this page).
-
-### Pull Requests
-
-All pull requests should be made to the `dev` branch of Bioscrape. To maintain code readability and validity, we encourage you to document your pull requests using the following ways:
-* Add a detailed comment when creating the pull request that summarizes the changes, features and/or bugs fixed.
-* We recommend that all new functions and classes have [docstrings](https://www.python.org/dev/peps/pep-0257/) and other documentation, if needed.
-* If possible, we encourage you to add test functions in the Tests directory that validate the code contributions. 
-* If your pull request is adding a new feature to the package, we also highly recommend a jupyter notebook example that goes along with that feature that discusses the use case. 
-
-## Styleguides
-
-* Following the PEP8 guideline, limit the first line to 72 characters or less
-* Reference issues and pull requests in your pull request comment 
-
-## Have any questions?
-
-If you have questions or would like to connect to the Bioscrape team on a regular basis, you can join our Slack channel. Bioscrape is a channel under the Synthetic Biology Modeling and Analysis Tools (SBTools) slack team.
-
-* [Join the SBTools Slack](https://join.slack.com/t/sbtools/shared_invite/zt-g82qjmvm-GAsNFLjyXGPlRBapqGDgFg)
-    * Use the `#bioscrape` channel for general questions or discussion about Bioscrape
-    * Use the `#general` channel for general questions about SBTools
-    * There are many other channels available for other synthetic biology modeling and analysis tools, check the channel list
