### 2023-01-23 biocircuits/bioscrape#157 @Robaina
JOSS REVIEW - Statement of Need -> Target Audience
Hi!

Me again, reviewing this [JOSS submissiojn](https://github.com/openjournals/joss-reviews/issues/5057).  Following JOSS guidelines, I see that you have included a nice Statement of Need in the manuscript. It clearly covers all points requested. However, there is a final point "Target audience" that I think could be stated explicitly in the manuscript,  i.e. bioscrape i built for researchers in "field, perhaps also educational?

Thanks!
