### 2023-01-23 #5057 @Robaina
Hi @sclamons,

I now have completed the review of this submission. I have raised some issues in the bioscrape repo (all linked here). Authors are already taking care of two of them (I just raised the other two).

I have locally installed bioscrape and tested all provided examples. I find the package a nice addition to the community, it is well documented and easy to use (congratulations to  the authors!)

There is only one mark that I left unchecked in the Review Checklist -> Performance. In the manuscript: Statement of Need, authors claim that bioscrape runs faster than competing methods (here probably I would also clarify in which functions it runs faster, i.e., parameter estimation?).  I think that such a claim should be backed by a benchmark analysis, which I did not see in the manuscript or the repo.
