### 2023-01-24 #5057 @Robaina
Hi @ayush9pandey, sorry for not tagging you. I wanted to address that message to both of you. Thanks, the repo and paper look very nice.

There is one minor thing I forgot to mention. The code you use in the example within "Statement of Need" in the paper is not runable. I guess because it corresponds to an older version of bioscrape. This is not that important in my opinion but just to let you know in case you want to update it.
