### 2023-01-24 #5057 @ayush9pandey
Thanks for bringing that up @Robaina ! We want to ensure backward compatibility as much as possible, so will ensure that the example in the paper works. I believe it is just a minor typo in the code in the figure.
