### 2023-01-26 817db2a
upgrade setuptools in deploy_bioscrape
