+++ .github/workflows/deploy_bioscrape.yml
@@ -21,6 +21,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
+        pip install --upgrade setuptools pip
         pip install -r requirements.txt
     - name: install bioscrape and lineage
       run: python setup.py install
