+++ paper.md
@@ -67,7 +67,7 @@ Once a given model is fully specified, it is then important to validate the mode
 
 The result is that a very large amount of data is needed to first parameterize and then validate models. The use of technologies for lab automation makes this data collection increasingly accessible and economical. For deterministic models, this may include data collected at many different operating conditions which can be achieved with high throughput measurement techniques involving liquid handling automation [@freemont_echo]. For stochastic models this may include large sample sizes of single cell measurements such as flow cytometry [@sachs_causal_2005;@zechner2012moment] and tracking single cell lineages with fluorescent microscopy [@kretzschmar2012lineage]. The Python API, simulation tools, and lineage module in bioscrape provide an ideal platform for such applications.
 
-Some popular software packages that do somewhat similar tasks as the bioscrape package are MATLAB's SimBiology toolbox [@MATLAB_2016], Stochpy [@stochpy], COPASI [@copasi], and Tellurium [@tellurium]. However, the bioscrape package is faster, supports fully general propensity functions, provides easy-to-use parameter identification interfaces, and allows more kinds of simulation than these alternatives making it more flexible and more efficient than alternative packages.
+Some popular software packages that do somewhat similar tasks as the bioscrape package are MATLAB's SimBiology toolbox [@MATLAB_2016], Stochpy [@stochpy], COPASI [@copasi], and Tellurium [@tellurium]. The SBML simulator libRoadRunner [@libroadrunner1;@libroadrunner2] is the state-of-the-art in deterministic and stochastic simulations of SBML models. Bioscrape simulation performance is of the same order as libRoadRunner and only around 20-30% slower while being an order of magnitude faster than other GUI-based simulators such as MATLAB and COPASI. However, the bioscrape package provides features beyond SBML simulations as it supports fully general propensity functions, provides easy-to-use parameter identification interfaces, and allows simulation of delays, and cell populations. The target audience for bioscrape includes researchers from diverse fields such as systems biology, synthetic biology, and chemical engineering. It is also aimed as an educational tool for classes on mathematical and computational biology.
 
 # Summary of features
 
