+++ paper.bib
@@ -18,6 +18,28 @@
   year={2006},
   publisher={Oxford University Press}
 }
+@article{libroadrunner2,
+  title={libRoadRunner 2.0: a high performance SBML simulation and analysis library},
+  author={Welsh, Ciaran and Xu, Jin and Smith, Lucian and K{\"o}nig, Matthias and Choi, Kiri and Sauro, Herbert M},
+  journal={Bioinformatics},
+  volume={39},
+  number={1},
+  pages={btac770},
+  year={2023},
+  publisher={Oxford University Press}
+}
+@article{libroadrunner1,
+  title={libRoadRunner: a high performance SBML simulation and analysis library},
+  author={Somogyi, Endre T and Bouteiller, Jean-Marie and Glazier, James A and K{\"o}nig, Matthias and Medley, J Kyle and Swat, Maciej H and Sauro, Herbert M},
+  journal={Bioinformatics},
+  volume={31},
+  number={20},
+  pages={3315--3321},
+  year={2015},
+  publisher={Oxford University Press}
+}
+
+
 @article{tellurium,
   title={Tellurium notebooks—an environment for reproducible dynamical modeling in systems biology},
   author={Medley, J Kyle and Choi, Kiri and K{\"o}nig, Matthias and Smith, Lucian and Gu, Stanley and Hellerstein, Joseph and Sealfon, Stuart C and Sauro, Herbert M},
