### 2023-01-27 biocircuits/bioscrape#157 @ayush9pandey
Closing the issue as joss-paper branch will automatically update the paper in the review issue: https://github.com/openjournals/joss-reviews/issues/5057

The branch will not be merged to bioscrape master.
