### 2023-01-27 #5057 @Robaina
> Thanks @Robaina ! I assume you wanted to address that message to me and not Sam. So I am replying to confirm that I am working on the issues that you raised on the repository. For the performance claims on the paper, I totally agree with you and will make required edits to the paper.

Right, I messed up the tags, sorry, corrected now.
