### 2023-01-27 #5057 @Robaina
Hi @csoneson ,

I now have checked all marks. Authors have addressed all the issues I raised (including the benchmark test on performance) and I don't have any more issues with bioscrape. I think it's ready for publication.

Thanks for inviting me to review this submission.
