### 2023-01-31 aa5eaf7
remove setup.py install for pip install in CI
