+++ README.md
@@ -99,8 +99,8 @@ will help you make the PR if you are new to `git` (just ask on the issue). More 
 
 Bioscrape versions:
 
-* 1.2.0 (latest release): To install run `pip install bioscrape==1.2.0` 
-* 1.1.0 (latest stable release): To install run `pip install bioscrape`
+* 1.2.1 (latest stable release): To install run `pip install bioscrape` 
+* 1.2.0 (tagged stable release): To install run `pip install bioscrape==1.2.0`
 * 1.0.4 (beta release): To install run `pip install bioscrape==1.0.4`
 
 # License
