### 2023-01-31 4b638df
remove setup.py install for pip install in CI
