+++ .github/workflows/deploy_bioscrape.yml
@@ -21,11 +21,10 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install --upgrade setuptools pip
         pip install -r requirements.txt
     - name: install bioscrape and lineage
-      run: python setup.py install
+      run: pip install .
     - name: Test bioscrape with pytest
       run: pytest
-    - name: install lineage
-      run: python setup.py install lineage
+    # - name: install lineage
+    #   run: python setup.py install lineage
