+++ .github/workflows/deploy_bioscrape.yml
@@ -23,7 +23,7 @@ jobs:
         python -m pip install --upgrade pip
         pip install -r requirements.txt
     - name: install bioscrape and lineage
-      run: pip install .
+      run: python setup.py install
     - name: Test bioscrape with pytest
       run: pytest
     - name: install lineage
