### 2023-01-31 50126ec
pip install instead of setup.py install works
