+++ .github/workflows/deploy_bioscrape.yml
@@ -23,7 +23,7 @@ jobs:
         python -m pip install --upgrade pip
         pip install -r requirements.txt
     - name: install bioscrape and lineage
-      run: python setup.py install
+      run: pip install .
     - name: Test bioscrape with pytest
       run: pytest
     - name: install lineage
