### 2023-01-31 faba33e
use pip install in test_bioscrape too
