### 2023-01-31 1ace5ef
fix readme.rst extension to .md in manifest
