### 2023-01-31 acb4734
Changes to package after JOSS review 1 (#160)
Updates in version 1.2.1:
* Fixes advanced examples notebook (Fix advanced examples notebook #154)
* Adds contributing.md (Add contributing.md  #158)
* Adds bioscrape simulation comparison (Benchmarking bioscrape performance #159)
* Installation via python setup.py install replaced with pip install.
* New contributor guidelines.
