### 2023-01-31 biocircuits/bioscrape#151 @ayush9pandey
Fixed and merged to master with #160 (available as bioscrape v1.2.1)
