### 2023-02-04 #5057 @Farnazmdi
Hi @csoneson  and @ayush9pandey
I finished my review of bioscrape. I believe it is a nice package with useful capabilities to people in the field. I left a few minor comments in the issues with "JOSS Review -- ..." in the titles. As soon as those are addressed I think it is ready!

Thank you for inviting me to review this package!
