### 2023-02-12 8d791f8
Add temperature to inference
