### 2023-02-19 7dd79d3
add docstring to py_inference
