### 2023-02-19 4e1f9c2
add docstrings to model and simulator UI functions
