### 2023-02-19 e66771d
add labels to figures in examples and update version to 1.2.3
