### 2023-02-19 d9e72b5
add labels to figures in examples and update version to 1.2.3
