### 2023-02-20 8ad4b0c
change pip install to python setup.py install
