### 2023-02-20 fc7938c
update CI to build with pyproject
