### 2023-02-20 0688c4e
add setuptools to requirements for testing
