+++ requirements.txt
@@ -1,3 +1,4 @@
+setuptools>=64
 matplotlib
 pytest
 numpy>=1.16.5
