### 2023-02-20 380543a
move back to python setup.py install
