+++ requirements.txt
@@ -1,4 +1,3 @@
-setuptools>=64
 matplotlib
 pytest
 numpy>=1.16.5
