### 2023-02-20 a91fb94
move back to python setup.py install
