### 2023-02-21 1d59eaa
add src/ to bioscrape to avoid users/CI to import package directly
