+++ pyproject.toml
@@ -4,3 +4,50 @@ build-backend = "setuptools.build_meta"
 
 [tool.distutils.bdist_wheel]
 universal = true
+
+[project]
+name = "Bioscrape"
+version = "1.2.2"
+description = "Biological Stochastic Simulation of Single Cell Reactions and Parameter Estimation"
+authors = [{name='Ayush Pandey', email='apandey@caltech.edu'},
+           {name='William Poole'},
+           {name='Anandh Swaminathan'}]
+classifiers=[
+        'Development Status :: 5 - Production/Stable',
+        'Intended Audience :: Science/Research',
+        'License :: OSI Approved :: BSD License',
+        'Programming Language :: Python :: 3.6',
+        'Topic :: Software Development',
+        'Topic :: Scientific/Engineering',
+        'Operating System :: OS Independent',
+    ]
+license = {file = "LICENSE"}
+readme = "README.md"
+keywords=["SBML", "synthetic biology", "modeling",
+          "Chemical Reaction Network", "CRN simulator", "stochastic",
+          "parameter inference"]
+dependencies = [
+    "numpy",
+    "matplotlib",
+    "pytest",
+    "scipy",
+    "cython",
+    "python-libsbml",
+    "beautifulsoup4",
+    "sympy",
+    "emcee",
+    "pandas"
+]
+requires-python = ">=3.6 <=3.11"
+
+[project.optional-dependencies]
+test = [
+  "pytest",
+]
+
+[project.urls]
+"Homepage" = "https://github.com/biocircuits/bioscrape/"
+"Bug Tracker" = "https://github.com/biocircuits/bioscrape/issues"
+
+[tool.setuptools]
+packages = ["bioscrape"]
