+++ .github/workflows/test_bioscrape.yml
@@ -22,7 +22,7 @@ jobs:
         python -m pip install --upgrade pip
         pip install -r requirements.txt
     - name: install bioscrape and lineage
-      run: python -m pip install . --wheel
+      run: pip install .
     - name: Test bioscrape with pytest
       run: pytest
     - name: install lineage
