+++ setup.py
@@ -1,0 +1,160 @@
+from distutils.core import setup
+from distutils.extension import Extension
+import platform
+import os
+import sys
+import subprocess
+
+# Set to true to enable line profiling
+line_debug = False
+
+#The following try-catch statements ensure numpy and cython are installed prior to running setup.py
+#in some virtual environments, pip dependencies have issues with cython packages.
+#NUMPY CHECK
+# try:
+#     from numpy import get_include
+# except ModuleNotFoundError:
+#     subprocess.check_call([sys.executable, "-m", "pip", "install", "numpy"])
+#     from numpy import get_include
+# #CYTHON CHECK
+# try:
+#     from Cython.Build import cythonize
+# except ModuleNotFoundError:
+#     subprocess.check_call([sys.executable, "-m", "pip", "install", "cython"])
+#     from Cython.Build import cythonize
+
+
+#Load the readme as a long description
+with open('README.md') as fp:
+    long_description = fp.read()
+
+# Compile Cython
+try:
+    numpyInclude = [get_include(), '.']
+
+    #Install Bioscrape Core Package
+    package_data = {'bioscrape': ['*.pxd']}
+    bioscrape_src_dir = 'bioscrape'
+
+    ext_options = {}
+    ext_options['language'] = 'c++'
+    ext_options['include_dirs'] = numpyInclude
+    if platform.system() == "Darwin":
+        ext_options['extra_compile_args'] = ['-std=c++11', "-mmacosx-version-min=10.9"]
+        ext_options['extra_link_args'] = ["-stdlib=libc++", "-mmacosx-version-min=10.9"]
+        print('Using macOS clang args')
+    if line_debug:
+        ext_options['define_macros'] = [('CYTHON_TRACE_NOGIL', '1')]
+
+    #used to generate HTML annotations of the cython code for
+    #optimization purposes.
+    cythonize_options = {
+    "include_path":[bioscrape_src_dir],
+    "language_level":"2" #Language level 3 does not work yet
+    }
+    if "annotate" in sys.argv:
+        cythonize_options['annotate'] = True
+        sys.argv.remove("annotate")
+    if line_debug:
+        cythonize_options["compiler_directives"] = {
+            'profile': True,
+            'linetrace': True,
+            'binding': True
+        }
+
+    # # Turn on to enable gdb debugging
+    # cythonize_options["gdb_debug"] = True
+
+    #Determine if we install bioscrape, lineage, or both
+    install_bioscrape = False
+    install_lineage = False
+    if "bioscrape" not in sys.argv and "lineage" not in sys.argv:
+        install_bioscrape = True
+        install_lineage = True
+    if "bioscrape" in sys.argv:
+        install_bioscrape = True
+        sys.argv.remove("bioscrape")
+    if "lineage" in sys.argv:
+        install_lineage = True
+        sys.argv.remove("lineage")
+
+    elif "bioscrape" not in sys.argv:
+        install_lineage = True
+
+    cython_extensions = []
+    if install_bioscrape:
+        print("Installing Bioscrape...")
+        bioscrape_source_files = ['random.pyx', 'types.pyx', 'simulator.pyx', 'inference.pyx']
+        bioscrape_extensions = [
+                Extension(
+                    name = 'bioscrape.'+s.split('.')[0],
+                    sources = [bioscrape_src_dir+'/'+s],
+                    **ext_options) for s in bioscrape_source_files
+            ]
+        cython_extensions += cythonize(bioscrape_extensions, **cythonize_options)
+        print("Bioscrape Cythonized.")
+
+    if install_lineage:
+        package_data['lineage'] = ['*.pxd']
[…]
