### 2023-02-23 97d0d0e
working pyproject setup with tests
