### 2023-02-23 12d2fd2
working pyproject setup with tests
