+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=64", "cython", "wheel"]
+requires = ["setuptools>=64", "Cython", "numpy", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [tool.distutils.bdist_wheel]
@@ -16,7 +16,13 @@ classifiers=[
         'Development Status :: 5 - Production/Stable',
         'Intended Audience :: Science/Research',
         'License :: OSI Approved :: BSD License',
-        'Programming Language :: Python :: 3.6',
+        'Programming Language :: Cython',
+        'Programming Language :: Python',
+        'Programming Language :: Python :: 3',
+        'Programming Language :: Python :: 3.7',
+        'Programming Language :: Python :: 3.8',
+        'Programming Language :: Python :: 3.9',
+        'Programming Language :: Python :: 3.10',
         'Topic :: Software Development',
         'Topic :: Scientific/Engineering',
         'Operating System :: OS Independent',
@@ -36,7 +42,9 @@ dependencies = [
     "beautifulsoup4",
     "sympy",
     "emcee",
-    "pandas"
+    "pandas",
+    "corner",
+    "lmfit"
 ]
 requires-python = ">=3.7"
 
@@ -51,5 +59,8 @@ test = [
 
 [tool.setuptools]
 packages = ["bioscrape"]
-# include-package-data = true
-package-dir = {"bioscrape"="src/bioscrape"}
+include-package-data = true
+package-dir = {"bioscrape"="bioscrape"}
+
+[tool.setuptools.exclude-package-data]
+"*" = ["*.c", "*.h"]
