+++ pip-wheel-metadata/Bioscrape-1.2.2.dist-info/top_level.txt
@@ -1,1 +1,0 @@
-bioscrape
