### 2023-02-25 #5057 @csoneson
@ayush9pandey - I went through the submission and it looks good; I made a small [PR](https://github.com/biocircuits/bioscrape/pull/167) with some fixes to the bibliography, please check and merge if you agree. You can then generate a new proof with `@editorialbot generate pdf`.

Then, the next steps are as follows:

- [ ] Make a tagged release of the software, and list the version tag of the archived version here.
- [ ] Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.
