### 2023-02-25 #5057 @ayush9pandey
Thanks @csoneson ! I merged the PR and tagged the release for v1.2.2 on Github. I submitted the archive to Zenodo: 10.5281/zenodo.7677726 https://zenodo.org/record/7677726
