### 2023-02-26 #5057 @csoneson
Thanks @ayush9pandey. It looks like there's still one DOI missing (for COPASI), which is likely (I notice now) because there are two entries in the .bib file for this paper (and only one has the DOI). Could you take a look and clean up the .bib file to only contain the references that are actually cited in the paper, and remove any duplicated entries? Thanks!
