+++ paper.bib
@@ -41,8 +41,6 @@
   publisher={Oxford University Press},
   doi={10.1093/bioinformatics/btv363}
 }
-
-
 @article{tellurium,
   title={Tellurium notebooks—an environment for reproducible dynamical modeling in systems biology},
   author={Medley, J Kyle and Choi, Kiri and K{\"o}nig, Matthias and Smith, Lucian and Gu, Stanley and Hellerstein, Joseph and Sealfon, Stuart C and Sauro, Herbert M},
@@ -54,26 +52,10 @@
   publisher={Public Library of Science San Francisco, CA USA},
   doi={10.1371/journal.pcbi.1006220}
 }
-@article{gunawardena2003chemical,
-  title={Chemical reaction network theory for in-silico biologists},
-  author={Gunawardena, Jeremy},
-  journal={Notes available for download at http://vcp. med. harvard. edu/papers/crnt. pdf},
-  year={2003}
-}
-@article{copasi,
-  title={COPASI—a complex pathway simulator},
-  author={Hoops, Stefan and Sahle, Sven and Gauges, Ralph and Lee, Christine and Pahle, J{\"u}rgen and Simus, Natalia and Singhal, Mudita and Xu, Liang and Mendes, Pedro and Kummer, Ursula},
-  journal={Bioinformatics},
-  volume={22},
-  number={24},
-  pages={3067--3074},
-  year={2006},
-  publisher={Oxford University Press}
-}
 @misc{github_bioscrape,
-  author = {Swaminathan, Anandh and Poole, William and Pandey, Ayush and Hsiao, Victoria and Murray, Richard M},
+  author = {Pandey, Ayush and Poole, William and Swaminathan, Anandh and Hsiao, Victoria and Murray, Richard M},
   title = {Bioscrape: Biological Stochastic Simulation of Single Cell Reactions and Parameter Estimation},
-  year = {2022},
+  year = {2023},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/biocircuits/bioscrape/}
@@ -111,71 +93,6 @@
   year={2012},
   publisher={IEEE Computer Society Press}
 }
-
-@article{atkinson2003development,
-  title={Development of genetic circuitry exhibiting toggle switch or oscillatory behavior in Escherichia coli},
-  author={Atkinson, Mariette R and Savageau, Michael A and Myers, Jesse T and Ninfa, Alexander J},
-  journal={Cell},
-  volume={113},
-  number={5},
-  pages={597--607},
-  year={2003},
-  publisher={Elsevier}
-}
-
-@article{siegal2014resource,
-  title={Resource usage and gene circuit performance characterization in a cell-free breadboard},
-  author={Siegal-Gaskins, Dan and Tuza, Zoltan A and Kim, Jongmin and Noireaux, Vincent and Murray, Richard M},
-  journal={BioRxiv},
-  doi={10.1101/000885},
-  pages={000885},
-  year={2014},
-  publisher={Cold Spring Harbor Laboratory}
-}
-
-@book{alon2006introduction,
-  title={An introduction to systems biology: design principles of biological circuits},
-  author={Alon, Uri},
-  year={2006},
-  publisher={Chapman and Hall/CRC}
-}
-
-@article{kondo2010reaction,
-  title={Reaction-diffusion model as a framework for understanding biological pattern formation},
-  author={Kondo, Shigeru and Miura, Takashi},
-  journal={Science},
-  doi={10.1126/science.1179047},
-  volume={329},
-  number={5999},
-  pages={1616--1620},
-  year={2010},
[…]
